컴포넌트 리팩토링: 이벤트 관리 UI 구성 요소 추가
- App.tsx에서 이벤트 관리 UI를 구성하는 컴포넌트를 분리하여 EventFormSection, CalendarViewSection, EventListSection으로 나누어 가독성을 향상
- 각 컴포넌트는 이벤트 추가/수정, 달력 뷰, 이벤트 목록 표시 기능을 담당
- 불필요한 코드 제거 및 구조 개선으로 유지보수 용이성 증가
